AscentAcademy_Event_Page
@AscentAcademy
Those who want to participate in the event can register on this page using the register button. In the 
header section,  the key speakers of the event, the schedule, various events, and so 
on; links for each of these sections. details of the speakers, 
venues, and the description of the event hosting.  page looks attractive; 
suitable background colors, font styles, and font colors. HTML and CSS are the primary tools you 
need to build this page